+++ HISTORY.rst
@@ -1,9 +1,23 @@
 .. :changelog:
 
-History
--------
+===============
+Release History
+===============
+
+0.2.0 (2015-08-30)
+------------------
+
+* Removed ``logger`` argument from :py:meth:`configloader.ConfigLoader.__init__`.
+* Improved test coverage.
+
+
+0.1.1 (2015-07-29)
+------------------
+
+* Made AttrDict and PyYAML optional dependencies.
+
 
 0.1.0 (2015-07-29)
----------------------
+------------------
 
 * First release on PyPI.